refactor simulation component
