+++ src/app/core/models/simulator.interface.ts
@@ -1,0 +1,10 @@
+export interface Niveles {
+    value: string;
+    viewValue: string;
+}
+
+export interface ChartData {
+    x: string;
+    y: number[];
+    label: string;
+}
